requirements.txt
azure-storage-blob==12.17.0
azure-functions==1.17.0
pandas==2.1.0
scikit-learn==1.3.2
mlflow==2.8.1
